# openapi.yaml
openapi: 3.1.0
info:
  title: Model Train Control System API
  description: |
    REST API for managing model trains, edge controllers, and hardware plugins.

    ## Architecture
    - **Central API**: FastAPI backend for train control and configuration
    - **Edge Controllers**: Raspberry Pi devices controlling physical trains
    - **MQTT**: Message broker for real-time command/telemetry

    ## Data Flow
    1. Commands sent via REST API → Published to MQTT
    2. Edge controllers subscribe to MQTT commands
    3. Telemetry from edge controllers → Stored in database
    4. Status retrieved via REST API from database
  version: 0.1.0
  contact:
    name: Model Train Control System
    url: https://github.com/bunchc/model-train-control-system
servers:
  - url: http://localhost:8000
    description: Local development server
  - url: http://central-api:8000
    description: Docker container deployment
paths:
...
  /api/trains:
    get:
      summary: List All Trains
      description: |
        Returns a list of all trains currently configured in the system.
        Note: Currently returns mock data. Will be replaced with database queries.
      operationId: list_trains
      tags:
        - trains
      responses:
        '200':
          description: Successful response with list of trains
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: '#/components/schemas/Train'
              example:
                - id: "1"
                  name: "Express"
                  description: "High-speed passenger train"
                  model: "Bachmann HO Scale"
                  plugin:
                    name: "motor_controller"
                    config:
                      i2c_address: "0x60"
                - id: "2"
                  name: "Freight"
                  description: "Heavy freight locomotive"
                  model: "Athearn Genesis"
                  plugin:
                    name: "motor_controller"
                    config:
                      i2c_address: "0x61"

    post:
      summary: Register New Train
      description: |
        Register a new train in the system. This endpoint accepts controller_uuid
        in the request body. Edge controllers should prefer using
        POST /api/controllers/{controller_uuid}/trains for clearer ownership.
      operationId: register_train
      tags:
        - trains
      requestBody:
        required: true
        description: Train registration data
        content:
          application/json:
            schema:
              type: object
              required:
                - train_id
                - name
                - controller_uuid
                - plugin
              properties:
                train_id:
                  type: string
                  description: Unique UUID for the train
                  example: "9bf2f703-5ba2-5032-a749-01cce962bcf6"
                name:
                  type: string
                  description: Human-readable train name
                  example: "Express Line Engine"
                controller_uuid:
                  type: string
                  description: UUID of the edge controller managing this train
                  example: "58902c61-2b70-42fe-8542-6b5b404aa998"
                description:
                  type: string
                  description: Optional train description
                  example: "DC Motor on M1 port"
                model:
                  type: string
                  description: Train model name/number
                  example: "DC Motor"
                plugin:
                  type: object
                  required:
                    - name
                  properties:
                    name:
                      type: string
                      example: "dc_motor"
                    config:
                      type: object
                      example:
                        motor_port: 1
      responses:
        '201':
          description: Train registered successfully
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Train'
        '400':
          description: Controller not found
        '409':
          description: Train ID already exists
...
  /api/trains/{train_id}/status:
    get:
      summary: Get Train Status from Database
      description: |
        Retrieves the latest status for a train from the database.
        Status data is periodically updated by edge controllers via MQTT.
      operationId: get_train_status
      tags:
        - trains
      parameters:
        - name: train_id
          in: path
          required: true
          description: Unique identifier of the train
          schema:
            type: string
          example: "1"
      responses:
        '200':
          description: Train status retrieved successfully
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/TrainStatus'
        '404':
          description: Train status not available
          content:
            application/json:
              schema:
                type: object
                properties:
                  detail:
                    type: string
                    example: "Train status not available"

  # ============================================================================
  # Configuration Endpoints
  # ============================================================================
...
  /api/config/trains/{train_id}:
    get:
      summary: Get Train Configuration by ID
      description: Get configuration details for a specific train
      operationId: get_train_config
      tags:
        - config
      parameters:
        - name: train_id
          in: path
          required: true
          description: Unique train identifier
          schema:
            type: string
          example: "train-001"
      responses:
        '200':
          description: Train configuration retrieved
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Train'
        '404':
          description: Train not found
          content:
            application/json:
              schema:
                type: object
                properties:
                  detail:
                    type: string
                    example: "Train not found"
...
  /api/config/edge-controllers/{edge_controller_id}:
    get:
      summary: Get Edge Controller Configuration
      description: Get configuration details for a specific edge controller
      operationId: get_edge_controller_config
      tags:
        - config
      parameters:
        - name: edge_controller_id
          in: path
          required: true
          description: Unique edge controller identifier
          schema:
            type: string
          example: "ec-001"
      responses:
        '200':
          description: Edge controller configuration retrieved
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/EdgeController'
        '404':
          description: Edge controller not found
          content:
            application/json:
              schema:
                type: object
                properties:
                  detail:
                    type: string
                    example: "Edge controller not found"
  /api/controllers/{controller_uuid}/trains:
    post:
      summary: Register Train for Controller
      description: |
        Register a new train managed by a specific edge controller.
        This is the preferred endpoint for edge controllers to self-register
        trains during initialization. The controller UUID must already exist.
      operationId: register_train_for_controller
      tags:
        - controllers
        - trains
      parameters:
        - name: controller_uuid
          in: path
          required: true
          description: UUID of the edge controller
          schema:
            type: string
          example: "58902c61-2b70-42fe-8542-6b5b404aa998"
      requestBody:
        required: true
        description: Train registration data
        content:
          application/json:
            schema:
              type: object
              required:
                - train_id
                - name
                - plugin
              properties:
                train_id:
                  type: string
                  description: Unique UUID for the train
                  example: "9bf2f703-5ba2-5032-a749-01cce962bcf6"
                name:
                  type: string
                  description: Human-readable train name
                  example: "Express Line Engine"
                description:
                  type: string
                  description: Optional train description
                  example: "DC Motor on M1 port"
                model:
                  type: string
                  description: Train model name/number
                  example: "DC Motor"
                plugin:
                  type: object
                  required:
                    - name
                  properties:
                    name:
                      type: string
                      description: Plugin name (dc_motor, stepper_motor, etc)
                      example: "dc_motor"
                    config:
                      type: object
                      description: Plugin-specific configuration
                      example:
                        motor_port: 1
      responses:
        '201':
          description: Train registered successfully
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Train'
        '400':
          description: Controller not found
          content:
            application/json:
              schema:
                type: object
                properties:
                  detail:
                    type: string
                    example: "Controller 58902c61-2b70-42fe-8542-6b5b404aa998 not found"
        '409':
          description: Train already exists
          content:
            application/json:
              schema:
                type: object
                properties:
                  detail:
                    type: string
                    example: "Train 9bf2f703-5ba2-5032-a749-01cce962bcf6 already exists"
        '500':
          description: Server error
          content:
            application/json:
              schema:
                type: object
                properties:
                  detail:
                    type: string
                    example: "Failed to register train: Database error"
  /api/config/trains:
    post:
      summary: Register Train (OpenAPI Alias)
      description: |
        Alias for POST /api/trains to maintain consistency with
        GET /api/config/trains endpoint naming.
      operationId: register_train_config_alias
      tags:
        - configuration
        - trains
      requestBody:
        $ref: '#/components/requestBodies/TrainRegistration'
      responses:
        '201':
          description: Train registered successfully
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Train'
        '400':
          description: Controller not found
        '409':
          description: Train already exists

  # ============================================================================
  # Internal Endpoints (used by edge controllers)
  # ============================================================================
  /api/status/update:
    post:
      summary: Update Train Status (Internal)
      description: |
        Internal endpoint for updating train status in the database.
        Typically called by edge controllers or MQTT bridge service.
      operationId: update_train_status
      tags:
        - internal
      requestBody:
        required: true
        content:
          application/json:
            schema:
              type: object
              required:
                - train_id
                - speed
                - voltage
                - current
                - position
              properties:
                train_id:
                  type: string
                  example: "1"
                speed:
                  type: integer
                  minimum: 0
                  maximum: 100
                  example: 50
                voltage:
                  type: number
                  format: float
                  example: 12.3
                current:
                  type: number
                  format: float
                  example: 0.8
                position:
                  type: string
                  example: "section_A"
      responses:
        '200':
          description: Status updated successfully
          content:
            application/json:
              schema:
                type: object
                properties:
                  message:
                    type: string
                    example: "Status updated"
                  train_id:
                    type: string
                    example: "1"

# ============================================================================
# Schema Components (aligned with Pydantic models)
# ============================================================================
components:
  schemas:
...
    TrainPlugin:
      type: object
      required:
        - name
      description: Plugin configuration for a specific train
      properties:
        name:
          type: string
          minLength: 1
          description: Name of the plugin to use
          example: "motor_controller"
        config:
          type: object
          additionalProperties: true
          description: Train-specific plugin configuration
          default: {}
          example:
            i2c_address: "0x60"
            default_speed: 50

    # ========================================================================
    # Train Schemas
    # ========================================================================
    Train:
      type: object
      required:
        - id
        - name
        - plugin
      properties:
        id:
          type: string
          minLength: 1
          description: Unique train identifier
          example: "train-001"
        name:
          type: string
          minLength: 1
          description: Human-readable train name
          example: "Express Locomotive"
        description:
          type: string
          nullable: true
          description: Train description
          example: "High-speed passenger train"
        model:
          type: string
          nullable: true
          description: Train model/manufacturer
          example: "Bachmann HO Scale"
        plugin:
          $ref: '#/components/schemas/TrainPlugin'
    TrainStatus:
      type: object
      required:
        - train_id
        - speed
        - voltage
        - current
        - position
      description: Real-time train telemetry and status
      properties:
        train_id:
          type: string
          minLength: 1
          description: Train identifier
          example: "train-001"
        speed:
          type: integer
          minimum: 0
          maximum: 100
          description: Current speed (0-100)
          example: 50
        voltage:
          type: number
          format: float
          minimum: 0
          description: Battery/track voltage in volts
          example: 12.3
        current:
          type: number
          format: float
          minimum: 0
          description: Current draw in amperes
          example: 0.8
        position:
          type: string
          minLength: 1
          description: Current track position/section
          example: "section_A"
...